Split draksetup and simplify packaging of drakrun
