+++ drakrun/.dockerignore
@@ -1,0 +1,3 @@
+.drakrun.egg-info/
+.mypy_cache/
+build/
