+++ drakrun/MANIFEST.in
@@ -1,3 +1,3 @@
-include drakrun/cfg.template
-include drakrun/tools/*
+include drakrun/templates/*
 include drakrun/test/*
+include drakrun/utilities/draksetup/tools/*
