+++ drakrun/debian/control
@@ -8,5 +8,5 @@ Standards-Version: 3.9.5
 Package: drakrun
 Architecture: amd64
 Depends: tcpdump, genisoimage, qemu-utils, bridge-utils, dnsmasq
-Pre-Depends: dpkg (>= 1.16.1), {{PYTHON_PACKAGES}}, python3-distutils
+Pre-Depends: dpkg (>= 1.16.1), python3, python3-distutils
 Description: DRAKVUF Sandbox Worker (standalone)
