+++ drakrun/debian/drakrun.install
@@ -1,4 +1,4 @@
-drakrun/config.ini /etc/drakrun/
-drakrun/scripts/* /etc/drakrun/scripts/
-drakrun/hooks.txt /etc/drakrun/
-drakrun/systemd/*.service /lib/systemd/system/
+drakrun/templates/cfg.template /etc/drakrun/scripts/cfg.template
+drakrun/templates/config.dist.ini /etc/drakrun/config.ini
+drakrun/templates/hooks.dist.txt /etc/drakrun/hooks.txt
+drakrun/templates/drakrun@.service /lib/systemd/system/
