+++ drakrun/drakrun/utilities/draksetup/_config.py
@@ -1,0 +1,6 @@
+import configparser
+
+from drakrun.paths import DRAKRUN_CONFIG_PATH
+
+config = configparser.ConfigParser()
+config.read(DRAKRUN_CONFIG_PATH)
