+++ drakrun/mypy.ini
@@ -1,5 +1,5 @@
 [mypy]
-python_version = 3.8
+python_version = 3.7
 
 [mypy-IPython.*]
 ignore_missing_imports = True
