+++ drakrun/package/build.sh
@@ -5,23 +5,18 @@ OUTDIR=$(pwd)/out
 # Base image for building
 if [ "$BASEIMAGE" = "" ]
 then
-    BASEIMAGE=debian:buster
+    BASEIMAGE=debian-buster
 fi
 
-# Current user - to make current user owner of output .deb files.
-FILE_OWNER=$(id -u):$(id -g)
-
 echo "Building $BASEIMAGE based image into $OUTDIR"
-docker build -f drakrun/package/Dockerfile \
-             --build-arg IMAGE=$BASEIMAGE \
+docker build -f drakrun/package/$BASEIMAGE.Dockerfile \
              -t deb-build . && \
 docker run --rm \
-           -i \
            -v $OUTDIR:/out \
            deb-build \
-           sh -c "package/find-python.sh && dpkg-buildpackage -us -uc -b && cp ../drakrun*.deb /out && chown -R $FILE_OWNER /out"
+           sh -c "cp ../drakrun*.deb /out"
 
 if [ $? -ne 0 ]; then echo Failed to build package ; exit 1 ; fi
 
-echo Verifying package existence...
+echo Packages created:
 find $OUTDIR
