+++ drakrun/requirements.txt
@@ -1,4 +1,6 @@
 karton-core==5.2.0
+boto3==1.28.20
+redis==4.4.4
 pefile==2019.4.18
 # Use pdbparse from master branch
 # current release - 1.5 has troubles parsing some of the PDBs
